+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.6.0] - 2021-02-17
+- **Added** support to empty closed tag like "<${FooBar}>...</>"
+- **Fixed** closed tag using a placeholder component
+
 ## [3.5.2] - 2021-01-23
 - **Improved** `mount` method